Moved supreme robot code out
