+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/BBLibrary/Hardware/BBGyro.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Hardware;
+package org.firstinspires.ftc.teamcode.BBLibrary.Hardware;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.hardware.bosch.JustLoggingAccelerationIntegrator;
