+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/BBLibrary/Hardware/BBMotor.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Hardware;
+package org.firstinspires.ftc.teamcode.BBLibrary.Hardware;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
