+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsytems/Drivetrain.java
@@ -1,14 +1,9 @@
 package org.firstinspires.ftc.teamcode.Subsytems;
 
-import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.Hardware.BBGyro;
-import org.firstinspires.ftc.teamcode.Hardware.BBMotor;
-
-import java.util.Timer;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.locks.ReentrantLock;
+import org.firstinspires.ftc.teamcode.BBLibrary.Hardware.BBGyro;
+import org.firstinspires.ftc.teamcode.BBLibrary.Hardware.BBMotor;
 
 /**
  * Created by Josh on 11/6/2017.
