+++ build.gradle
@@ -6,6 +6,7 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.0.0'
@@ -13,3 +14,7 @@ buildscript {
 }
 
 
+
+repositories {
+    google()
+}
